requirements.txt
PyJWT
APScheduler==3.6.3
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
docopt==0.6.2
eight==1.0.0
Flask==1.1.2
Flask-APScheduler==1.11.0
Flask-Sockets==0.2.1
future==0.18.2
gevent==20.5.2
gevent-websocket==0.10.1
greenlet==0.4.16
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.5.2
MarkupSafe==1.1.1
pipreqs==0.4.10
pycparser==2.20
pycryptodomex==3.9.8
pyodbc==4.0.30
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
signxml==2.8.0
six==1.15.0
tzlocal==2.1
urllib3==1.25.10
waitress==1.4.3
Werkzeug==1.0.1
yarg==0.1.9
zope.event==4.4
zope.interface==5.1.0
